% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lat_prop_plot}
\alias{lat_prop_plot}
\title{Plot parameters from a dataframe vs latitude}
\usage{
lat_prop_plot(df, property, lat_range = NULL)
}
\arguments{
\item{df}{a dataframe with Dec_LAT}

\item{property}{Chemical property to plot}

\item{lat_range}{a two element vector of decimal degrees}
}
\value{
A ggplot object
}
\description{
Plot parameters from a dataframe vs latitude
}
