% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eims.R
\name{plot_eims_ratio}
\alias{plot_eims_ratio}
\title{Timeseries plot of mass ratio}
\usage{
plot_eims_ratio(eims_data, masses)
}
\arguments{
\item{eims_data}{A dataframe of EIMS data}

\item{masses}{A two element character vector of mass column names}
}
\value{
Timeseries plot
}
\description{
Timeseries plot of mass ratio
}
