% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underway.R
\name{read_en_underway}
\alias{read_en_underway}
\title{Read and parse Endeavor underway files}
\usage{
read_en_underway(ship_file)
}
\arguments{
\item{ship_file}{Path to an Endeavor format processed underway file.}
}
\value{
A dataframe of ship data
}
\description{
Read and parse Endeavor underway files
}
