% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underway.R
\name{read_underway}
\alias{read_underway}
\title{Read and parse underway files}
\usage{
read_underway(ship_file)
}
\arguments{
\item{ship_file}{Path to an Armstrong format processed underway file.}
}
\value{
A dataframe of ship data
}
\description{
Read and parse underway files
}
