% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eims.R
\name{plot_ts}
\alias{plot_ts}
\title{Timeseries plot of eims mass}
\usage{
plot_ts(data, field, timestamp = "timestamp")
}
\arguments{
\item{data}{A dataframe}

\item{field}{Character string of name of column.}

\item{timestamp}{Character string of name timestamp column.}
}
\value{
Timeseries plot
}
\description{
Timeseries plot of eims mass
}
